youtube-likes-automation/requirements.txt
google-api-python-client==2.159.0
google-auth-oauthlib==1.2.1
google-auth==2.37.0
requests==2.32.3
